build/publish fixes
pygments doesn't like graphql code-blocks :-(

enhance scripts so we avoid this mess again.

+++ utils/release.sh
@@ -6,5 +6,14 @@ if [ "$PIPENV_ACTIVE" = "1" ]; then
   exit 1
 fi
 
+SETUP_VERSION=`./setup.py --version`
+MOD_VERSION=`PYTHONPATH='.' python3 -c 'import sgqlc; print(sgqlc.__version__)'`
+if [ $SETUP_VERSION != $MOD_VERSION ]; then
+  echo "update both setup.py and sgqlc/__init__.py so versions match" >&2
+  exit 1
+fi
+
 rm -fr dist/ build
 ./setup.py egg_info -Db '' sdist bdist_wheel bdist_egg
+
+twine check dist/*.tar.gz || exit 1
